Fix logic to set the default timeout.
Although the default timeout is correctly initialized, it was overwritten
with random garbage when the LDAP2DNS_TIMEOUT environment variable
wasn't set.

Update the LDAP2DNS_RECLIMIT logic to be consistent with the handling
of LDAP2DNS_TIMEOUT.

Add help documentation for the "-t" option.

Remove a few more unused variables as detected by clang.