Add in-memory caching of data.  Cleanup.
Paramaterize a number of manifest constants with #define.

Expanded zone.location[] to 3 characters in order that it can always be
null-terminated and safely used as a regular C string.

Add in-memory caching of zones, location records and resource records.

Add debugging output of (some of) the in-memory cached items (under #define DEBUG only).